tests/wpt/tests/css/css-shapes/parsing/shape-margin-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Shapes Module Level 1: parsing shape-margin with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-shapes/#shape-margin-property">
<meta name="assert" content="shape-margin supports the full grammar '<length> | <percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("shape-margin", "0", "0px");
test_valid_value("shape-margin", "10px");
test_valid_value("shape-margin", "20em");
test_valid_value("shape-margin", "37.5%");
</script>
</body>
</html>
